[AIRFLOW-6721] Move Hive PrestoToMySqlTransfer test
Move the PrestoToMySqlTransfer test from the operators module
into an existing test module and use shared Hive test environment.
