+++ tests/providers/mysql/operators/test_presto_to_mysql.py
@@ -15,22 +15,23 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-
+import os
 import unittest
 from unittest.mock import patch
 
 from airflow.providers.mysql.operators.presto_to_mysql import PrestoToMySqlTransfer
+from tests.providers.apache.hive import DEFAULT_DATE, TestHiveEnvironment
 
 
-class TestPrestoToMySqlTransfer(unittest.TestCase):
+class TestPrestoToMySqlTransfer(TestHiveEnvironment):
 
     def setUp(self):
         self.kwargs = dict(
             sql='sql',
             mysql_table='mysql_table',
             task_id='test_presto_to_mysql_transfer',
-            dag=None
         )
+        super().setUp()
 
     @patch('airflow.providers.mysql.operators.presto_to_mysql.MySqlHook')
     @patch('airflow.providers.mysql.operators.presto_to_mysql.PrestoHook')
@@ -52,3 +53,20 @@ class TestPrestoToMySqlTransfer(unittest.TestCase):
         mock_mysql_hook.return_value.run.assert_called_once_with(self.kwargs['mysql_preoperator'])
         mock_mysql_hook.return_value.insert_rows.assert_called_once_with(
             table=self.kwargs['mysql_table'], rows=mock_presto_hook.return_value.get_records.return_value)
+
+    @unittest.skipIf(
+        'AIRFLOW_RUNALL_TESTS' not in os.environ,
+        "Skipped because AIRFLOW_RUNALL_TESTS is not set")
+    def test_presto_to_mysql(self):
+        op = PrestoToMySqlTransfer(
+            task_id='presto_to_mysql_check',
+            sql="""
+                SELECT name, count(*) as ccount
+                FROM airflow.static_babynames
+                GROUP BY name
+                """,
+            mysql_table='test_static_babynames',
+            mysql_preoperator='TRUNCATE TABLE test_static_babynames;',
+            dag=self.dag)
+        op.run(start_date=DEFAULT_DATE, end_date=DEFAULT_DATE,
+               ignore_ti_state=True)
